refactor: rename useLibraryBehavior module as behavior
